+++ service/ssm/api.go
@@ -14206,7 +14206,13 @@ func (c *SSM) UpdateAssociationRequest(input *UpdateAssociationInput) (req *requ
 //
 // Updates an association. You can update the association name and version,
 // the document version, schedule, parameters, and Amazon Simple Storage Service
-// (Amazon S3) output.
+// (Amazon S3) output. When you call UpdateAssociation, the system drops all
+// optional parameters from the request and overwrites the association with
+// null values for those parameters. This is by design. You must specify all
+// optional parameters in the call, even if you are not changing the parameters.
+// This includes the Name parameter. Before calling this API action, we recommend
+// that you call the DescribeAssociation API operation and make a note of all
+// optional parameters required for your UpdateAssociation call.
 //
 // In order to call this API operation, your Identity and Access Management
 // (IAM) user account, group, or role must be configured with permission to
@@ -14217,7 +14223,8 @@ func (c *SSM) UpdateAssociationRequest(input *UpdateAssociationInput) (req *requ
 // resource: <resource_arn>
 //
 // When you update an association, the association immediately runs against
-// the specified targets.
+// the specified targets. You can add the ApplyOnlyAtCronInterval parameter
+// to run the association during the next schedule run.
 //
 // Returns awserr.Error for service API and SDK errors. Use runtime type assertions
 // with awserr.Error's Code and Message methods to get detailed information about
@@ -16161,6 +16168,13 @@ type Association struct {
 	AssociationVersion *string `type:"string"`
 
 	// The version of the document used in the association.
+	//
+	// State Manager doesn't support running associations that use a new version
+	// of a document if that document is shared from another account. State Manager
+	// always runs the default version of a document if shared from another account,
+	// even though the Systems Manager console shows that a new version was processed.
+	// If you want to run an association using a new version of a document shared
+	// form another account, you must set the document version to default.
 	DocumentVersion *string `type:"string"`
 
 	// The managed node ID.
@@ -19268,7 +19282,7 @@ type Command struct {
 
 	// The number of targets for which the command invocation reached a terminal
 	// state. Terminal states include the following: Success, Failed, Execution
-	// Timed Out, Delivery Timed Out, Canceled, Terminated, or Undeliverable.
+	// Timed Out, Delivery Timed Out, Cancelled, Terminated, or Undeliverable.
 	CompletedCount *int64 `type:"integer"`
 
 	// The number of targets for which the status is Delivery Timed Out.
@@ -19370,8 +19384,8 @@ type Command struct {
 	//    more invocations doesn't have a value of Success but not enough invocations
 	//    failed for the status to be Failed. This is a terminal state.
 	//
-	//    * Canceled: The command was terminated before it was completed. This is
-	//    a terminal state.
+	//    * Cancelled: The command was terminated before it was completed. This
+	//    is a terminal state.
 	//
 	//    * Rate Exceeded: The number of managed nodes targeted by the command exceeded
 	//    the account limit for pending invocations. The system has canceled the
@@ -19748,8 +19762,8 @@ type CommandInvocation struct {
 	//    Invocation failures count against the MaxErrors limit of the parent command.
 	//    This is a terminal state.
 	//
-	//    * Canceled: The command was terminated before it was completed. This is
-	//    a terminal state.
+	//    * Cancelled: The command was terminated before it was completed. This
+	//    is a terminal state.
 	//
 	//    * Undeliverable: The command can't be delivered to the managed node. The
 	//    managed node might not exist or might not be responding. Undeliverable
@@ -19980,8 +19994,8 @@ type CommandPlugin struct {
 	//    Invocation failures count against the MaxErrors limit of the parent command.
 	//    This is a terminal state.
 	//
-	//    * Canceled: The command was terminated before it was completed. This is
-	//    a terminal state.
+	//    * Cancelled: The command was terminated before it was completed. This
+	//    is a terminal state.
 	//
 	//    * Undeliverable: The command can't be delivered to the managed node. The
 	//    managed node might not exist, or it might not be responding. Undeliverable
@@ -21207,6 +21221,13 @@ type CreateAssociationInput struct {
 
 	// The document version you want to associate with the target(s). Can be a specific
 	// version or the default version.
+	//
+	// State Manager doesn't support running associations that use a new version
+	// of a document if that document is shared from another account. State Manager
+	// always runs the default version of a document if shared from another account,
+	// even though the Systems Manager console shows that a new version was processed.
+	// If you want to run an association using a new version of a document shared
+	// form another account, you must set the document version to default.
 	DocumentVersion *string `type:"string"`
 
 	// The managed node ID.
@@ -28646,7 +28667,7 @@ func (s *DocumentDefaultVersionDescription) SetName(v string) *DocumentDefaultVe
 	return s
 }
 
-// Describes a Amazon Web Services Systems Manager document (SSM document).
+// Describes an Amazon Web Services Systems Manager document (SSM document).
 type DocumentDescription struct {
 	_ struct{} `type:"structure"`
 
@@ -28659,6 +28680,13 @@ type DocumentDescription struct {
 
 	// The user in your organization who created the document.
 	Author *string `type:"string"`
+
+	// The classification of a document to help you identify and categorize its
+	// use.
+	Category []*string `type:"list"`
+
+	// The value that identifies a document's category.
+	CategoryEnum []*string `type:"list"`
 
 	// The date when the document was created.
 	CreatedDate *time.Time `type:"timestamp"`
@@ -28707,7 +28735,7 @@ type DocumentDescription struct {
 	// The version of the document that is currently under review.
 	PendingReviewVersion *string `type:"string"`
 
-	// The list of OS platforms compatible with this SSM document.
+	// The list of operating system (OS) platforms compatible with this SSM document.
 	PlatformTypes []*string `type:"list"`
 
 	// A list of SSM documents required by a document. For example, an ApplicationConfiguration
@@ -28781,6 +28809,18 @@ func (s *DocumentDescription) SetAttachmentsInformation(v []*AttachmentInformati
 // SetAuthor sets the Author field's value.
 func (s *DocumentDescription) SetAuthor(v string) *DocumentDescription {
 	s.Author = &v
+	return s
+}
+
+// SetCategory sets the Category field's value.
+func (s *DocumentDescription) SetCategory(v []*string) *DocumentDescription {
+	s.Category = v
+	return s
+}
+
+// SetCategoryEnum sets the CategoryEnum field's value.
+func (s *DocumentDescription) SetCategoryEnum(v []*string) *DocumentDescription {
+	s.CategoryEnum = v
 	return s
 }
 
@@ -30833,8 +30873,8 @@ type GetCommandInvocationOutput struct {
 	//    wasn't zero. Invocation failures count against the MaxErrors limit of
 	//    the parent command. This is a terminal state.
 	//
-	//    * Canceled: The command was terminated before it was completed. This is
-	//    a terminal state.
+	//    * Cancelled: The command was terminated before it was completed. This
+	//    is a terminal state.
 	//
 	//    * Undeliverable: The command can't be delivered to the managed node. The
 	//    node might not exist or might not be responding. Undeliverable invocations
@@ -55014,6 +55054,13 @@ type UpdateAssociationInput struct {
 	ComplianceSeverity *string `type:"string" enum:"AssociationComplianceSeverity"`
 
 	// The document version you want update for the association.
+	//
+	// State Manager doesn't support running associations that use a new version
+	// of a document if that document is shared from another account. State Manager
+	// always runs the default version of a document if shared from another account,
+	// even though the Systems Manager console shows that a new version was processed.
+	// If you want to run an association using a new version of a document shared
+	// form another account, you must set the document version to default.
 	DocumentVersion *string `type:"string"`
 
 	// The maximum number of targets allowed to run the association at the same